requirements.txt
PyYAML==3.11
requests==2.10.0
twilio==5.4.0
